+++ build.gradle
@@ -199,40 +199,93 @@ repositories {
             includeGroup 'dev.ftb.mods'
         }
     }
+
+    maven { url = "https://maven.createmod.net" } // Create, Ponder, Flywheel
+    maven { url = "https://maven.ithundxr.dev/mirror" } // Registrate
+    maven { url = "https://raw.githubusercontent.com/Fuzss/modresources/main/maven/" } // ForgeConfigAPIPort
+
 }
 
 dependencies {
-    minecraft 'net.minecraftforge:forge:1.20.1-47.3.7'
+    minecraft 'net.minecraftforge:forge:1.20.1-47.4.6'
 
     annotationProcessor 'org.spongepowered:mixin:0.8.5:processor'
 
     compileOnly 'org.projectlombok:lombok:1.18.24'
     annotationProcessor 'org.projectlombok:lombok:1.18.24'
 
+
+    // ========================
+    // COMPILE MOD-ADDONS
+    // ========================
+
+    //JEI
     compileOnly(fg.deobf("mezz.jei:jei-${mc_version}-common-api:${jei_version}"))
     compileOnly(fg.deobf("mezz.jei:jei-${mc_version}-forge-api:${jei_version}"))
 
-    // at runtime, use the full JEI jar for Forge
-    runtimeOnly(fg.deobf("mezz.jei:jei-${mc_version}-forge:${jei_version}"))
-    implementation fg.deobf("curse.maven:jade-324717:${jade_id}")
+    //JADE (WAILA)
+    compileOnly fg.deobf("curse.maven:jade-324717:${jade_id}")
 
-    implementation fg.deobf("curse.maven:mekanism-268560:5125665")
-    implementation fg.deobf("curse.maven:pneumaticcraft-281849:5279325")
+    // MEKANISM
+    compileOnly fg.deobf("curse.maven:mekanism-268560:5125665")
 
-    implementation fg.deobf("dev.latvian.mods:kubejs-forge:${kubejs_version}")
-    implementation fg.deobf("dev.latvian.mods:rhino-forge:${rhino_version}")
-    implementation fg.deobf("dev.architectury:architectury-forge:${architectury_version}")
+    //PNEUMATICCRAFT (PNCR)
+    compileOnly fg.deobf("curse.maven:pneumaticcraft-281849:5279325")
 
-    runtimeOnly fg.deobf("curse.maven:probejs-585406:5227399")
+    //KUBEJS
+    compileOnly fg.deobf("dev.latvian.mods:kubejs-forge:${kubejs_version}")
+    compileOnly fg.deobf("dev.latvian.mods:rhino-forge:${rhino_version}")
+    compileOnly fg.deobf("dev.architectury:architectury-forge:${architectury_version}")
 
-    implementation fg.deobf("com.simibubi.create:create-${create_minecraft_version}:${create_version}:slim") { transitive = false }
-    implementation fg.deobf("com.jozufozu.flywheel:flywheel-forge-${flywheel_minecraft_version}:${flywheel_version}")
-    implementation fg.deobf("com.tterrag.registrate:Registrate:${registrate_version}")
+    // CREATE MOD
+    compileOnly(fg.deobf("com.simibubi.create:create-${mc_version}:${create_version}:slim") { transitive = false })
+    compileOnly(fg.deobf("net.createmod.ponder:Ponder-Forge-${mc_version}:${ponder_version}"))
+    compileOnly(fg.deobf("dev.engine-room.flywheel:flywheel-forge-api-${mc_version}:${flywheel_version}"))
+    compileOnly(fg.deobf("com.tterrag.registrate:Registrate:${registrate_version}"))
+    compileOnly(annotationProcessor("io.github.llamalad7:mixinextras-common:0.4.1"))
+    compileOnly("io.github.llamalad7:mixinextras-forge:0.4.1")
 
+    //BOTANIA
     compileOnly fg.deobf("vazkii.botania:Botania:${botania_version}")
 
-//    runtimeOnly fg.deobf("curse.maven:patchouli-306770:4892162")
-//    runtimeOnly fg.deobf("curse.maven:curios-309927:5367944")
+
+    // ========================
+    // RUNTIME MOD-ADDONS
+    // ========================
+
+    // PATCHUOLI
+    runtimeOnly fg.deobf("curse.maven:patchouli-306770:${patchouli_version}")
+
+    // CURIOS
+    runtimeOnly fg.deobf("curse.maven:curios-309927:${curios_version}")
+
+    // CREATE MOD
+    runtimeOnly(fg.deobf("com.simibubi.create:create-${mc_version}:${create_version}:slim") { transitive = false })
+    runtimeOnly(fg.deobf("net.createmod.ponder:Ponder-Forge-${mc_version}:${ponder_version}"))
+    runtimeOnly(fg.deobf("dev.engine-room.flywheel:flywheel-forge-${mc_version}:${flywheel_version}"))
+    runtimeOnly(fg.deobf("com.tterrag.registrate:Registrate:${registrate_version}"))
+    runtimeOnly("io.github.llamalad7:mixinextras-forge:0.4.1")
+
+    // ProbeJS (kubejs)
+    runtimeOnly fg.deobf("curse.maven:probejs-585406:5227399")
+
+    // JEI
+    runtimeOnly(fg.deobf("mezz.jei:jei-${mc_version}-forge:${jei_version}"))
+
+    //JADE (WAILA)
+    runtimeOnly fg.deobf("curse.maven:jade-324717:${jade_id}")
+
+    // MEKANISM
+    runtimeOnly fg.deobf("curse.maven:mekanism-268560:5125665")
+
+    //PNEUMATICCRAFT (PNCR)
+    runtimeOnly fg.deobf("curse.maven:pneumaticcraft-281849:5279325")
+
+    // KUBEJS
+    runtimeOnly fg.deobf("dev.latvian.mods:kubejs-forge:${kubejs_version}")
+    runtimeOnly fg.deobf("dev.latvian.mods:rhino-forge:${rhino_version}")
+    runtimeOnly fg.deobf("dev.architectury:architectury-forge:${architectury_version}")
+
 }
 
 // Example for how to get properties into the manifest for reading at runtime.
