+++ gradle.properties
@@ -15,11 +15,13 @@ kubejs_version=2001.6.5-build.16
 rhino_version=2001.2.3-build.10
 architectury_version=9.2.14
 
-create_minecraft_version = 1.20.1
-flywheel_minecraft_version = 1.20.1
-create_version = 0.5.1.f-26
-flywheel_version = 0.6.10-7
-registrate_version = MC1.20-1.3.3
+create_version=6.0.6-150
+flywheel_version=1.0.4
+registrate_version=MC1.20-1.3.3
+ponder_version=1.0.80
+
+patchouli_version=6164575
+curios_version=6418456
 
 botania_version = 1.20.1-450-FORGE
