+++ src/main/java/io/ticticboom/mods/mm/port/kinetic/CreateKineticPortType.java
@@ -1,7 +1,6 @@
 package io.ticticboom.mods.mm.port.kinetic;
 
-import com.simibubi.create.content.kinetics.BlockStressValues;
-import io.ticticboom.mods.mm.Ref;
+import com.simibubi.create.api.stress.BlockStressValues;
 import io.ticticboom.mods.mm.compat.kjs.builder.PortConfigBuilderJS;
 import io.ticticboom.mods.mm.model.PortModel;
 import io.ticticboom.mods.mm.port.IPortParser;
@@ -25,7 +24,8 @@ import java.util.function.Consumer;
 public class CreateKineticPortType extends PortType {
 
     public CreateKineticPortType() {
-        BlockStressValues.registerProvider(Ref.ID, new MMKineticStressProvider());
+        BlockStressValues.IMPACTS.registerProvider(MMKineticStressProviders::impactSupplier);
+        BlockStressValues.CAPACITIES.registerProvider(MMKineticStressProviders::capacitySupplier);
     }
 
     @Override
