+++ src/main/resources/META-INF/mods.toml
@@ -70,3 +70,9 @@ Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed mollis lacinia magn
     mandatory=false
     versionRange="[2001,2002)"
     ordering="BEFORE"
+[[dependencies.mm]]
+    modId="create"
+    mandatory=false
+    versionRange="[6.0.6,6.1.0)"
+    ordering="NONE"
+    side="BOTH"
